Different tool recommendations for OSX
Installed gems in OSX usually goes to a folder inside /Library, which
may get deleted and recreated after upgrading the operating system.

This changes proposes OSX users to use of homebrew and home folder
install, which will be more resilient to upgrades. Also, depending on
how homebrew was installed, this will also not require super-user
privileges.