Using Antora as a static file build
Since Asciidoctor does not support output to multiple pages, we need to
use some "static site builder" to achieve the same result.

One possible tool is Antora. This commit contains the minimal code
change to be able to compile the website with Antora. Core features
are well-documented and the end result is quite nice. There are some
caveats:

* They are strongly opiniated about directory structure. It did match
the old structure to a certain degree, though. Nonetheless, this helps
modularizing the pages.
* There is no obvious option to generate single-pages. This would impact
both single-page HTML and PDF outputs. In theory, this can be achieved
with a different generator, but that feature is not properly
documented, or manually creating composite index pages, which
requires some duplicate page referencing (one for nav, one for
composition). Since the user value for single-pages is an offline
version of the documentation, this can be achieved with a ZIP
distribution, built as a Github Action for each push or pull request.
* Setting up a building environment requires more steps. Either installation
of NodeJS plus Antora or usage of Docker. Both seems quite invasive if a
new dev does not have either already available. A properly setup
"package.json" file improves this experience, and the installation of
NodeJS is easier in Windows and Mac. This is quite important for Windows
contributors, since Ruby and Docker are not as trivial to install as
NodeJS.
* There is no catalog for different UIs. Antora's default output is
pretty, but requires some mandatory changes to be useful. This currently
requires a new repository.
* Integration with Travis and Github Actions are quite simple, since
NodeJS takes care of dependencies. It would requre some tests to confirm
if the old upload step still works.

See LibrePCB/librepcb-doc#20