Moving images to match Hugo standards
Originally, Asciidoc files were pointing to images using relative
folders. Hugo supports that, but the image folder needs to be a sibling
of the file referencing it.

Given this restriction, it seems creating multiple folders resolved the
issue, and it also helps with image organization, since the page
referencing a given image is easier to track.

This change does not fix image alignments. Unfortunatelly, some image
"flows" (such as the right/left floats) are still broken.