Integration of Hugo with Travis and Github Actions
This changes the Travis setup to install Hugo and the build script to
call Hugo, outputting to the same directory as the old script.

It also integrates with GitHub Actions, to allow contributors to provide
a ZIP file with a generated site for proof-reading.

Instruction in CONTRIBUTING and README were also updated to reflect it.