Using a different "profile" for offline builds
Hugo supports multiple configuration profiles, which can be leverage to
enable offline builds. This commit introduces an "offline" profile,
which adds a flag to build using relative URLs. This can be used to
build copies that can be read without a website.

Instructions for using this profile were included in the README.

This is also being used for the GitHub Action, to enable a proper
downloadable ZIP file as part of pull-requests.