Use wget for offline distribution
Hugo is amazing, but it relies on directory index (eg: /path/) instead
of using links to "index.html". If "relativeURLs" and "uglyURLs" are
enabled, HTMLs are generated just fine, but image links are not changed
to reflect the new link structure.

wget can mirror websites and update links automatically, this makes it
an easy workaround. Our offline copy will be as close as possible to
production and to the development setup.