Collapsable menus
This introduces a small Javascript snipped at the bottom of the
generated output. This JS block will hide all sections after the page
loads and show the current visible anchor.

The script will also inject a click listener to keep only the currently
visited item open. Since the output is a single HTML file, this trick
won't be a nuisance.

This is transparent for readers without Javascript enabled. They are
presumably rare, but we never know. Those users will have to deal with a
long TOC.